Move _computeTotalDifficulty to BlockchainBase
